+++ src/auth/auth.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
+import { forwardRef } from '@nestjs/common/utils';
 
 // import local module
-import { UsersModule } from '../users/users.module';
+import { UsersModule } from '../models/users/users.module';
 
 // import controller
 import { AuthController } from './auth.controller';
@@ -12,8 +13,7 @@ import { AuthService } from './auth.service';
 
 // import JWT
 import { JwtModule } from '@nestjs/jwt';
-import { JwtStrategy } from '../strategy/jwt.strategy';
-import { forwardRef } from '@nestjs/common/utils';
+import { JwtStrategy } from './strategy/jwt.strategy';
 
 @Module({
   imports: [
@@ -23,7 +23,7 @@ import { forwardRef } from '@nestjs/common/utils';
       useFactory: async (config: ConfigService) => {
         return {
           secret: config.get('JWT_SECRET'),
-          signOptions: { expiresIn: '2h' },
+          signOptions: { expiresIn: 2 },
         };
       },
       inject: [ConfigService],
