+++ src/auth/auth.service.ts
@@ -8,7 +8,7 @@ import { JwtService } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
 
 // import local service
-import { UsersService } from '../users/users.service';
+import { UsersService } from '../models/users/users.service';
 
 // import prisma
 import { PrismaClient } from '@prisma/client';
@@ -27,7 +27,7 @@ import {
 import * as bcrypt from 'bcrypt';
 
 // import error codes
-import { prismaErrorCodes } from '../errorCode/prismaErrorCode.enum';
+import { prismaErrorCodes } from '../common/constants/prismaErrorCode.enum';
 
 @Injectable()
 export class AuthService {
