+++ src/auth/strategy/jwt.strategy.ts
@@ -2,7 +2,7 @@ import { ExtractJwt, Strategy } from 'passport-jwt';
 import { PassportStrategy } from '@nestjs/passport';
 import { Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { NguoiDungDto } from '../dto/index.dto';
+import { NguoiDungDto } from '../../dto/index.dto';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy, 'JwtAuth') {
