+++ src/common/decorators/roles.decorator.ts
@@ -1,5 +1,5 @@
 import { SetMetadata } from '@nestjs/common';
-import { LoaiNguoiDung } from '../dto/index.dto';
+import { LoaiNguoiDung } from '../../dto/index.dto';
 
 export const ROLES_KEY = 'roles';
 export const Roles = (...roles: LoaiNguoiDung[]) =>
