+++ src/common/exceptions/http-exceptions.filter.ts
