+++ src/common/guards/jwt.guard.ts
@@ -14,12 +14,12 @@ export class JwtAuthGuard extends AuthGuard('JwtAuth') {
     return super.canActivate(context);
   }
 
-  handleRequest(err, user, info) {
+  handleRequest(err, user, info: Error | null) {
     // You can throw an exception based on either "info" or "err" arguments
     if (err || !user) {
       throw new UnauthorizedException({
-        ...info,
-        statusCode: HttpStatus.UNAUTHORIZED,
+        message: 'Unauthorized. Token related',
+        error: info instanceof Error ? info.message : info,
       });
     }
     return user;
