+++ src/common/interceptors/transform-to-plain.interceptor.ts
