+++ src/common/pipes/parse-file.pipe.ts
