+++ src/dto/index.dto.ts
@@ -1,6 +1,6 @@
 export * from './common.dto';
 export * from './upload.dto';
-export * from '../users/user-dto/user.dto';
-export * from '../movie/movie-dto/movie.dto';
-export * from '../theatre/theatre-dto/theatre.dto';
-export * from '../booking/booking-dto/booking.dto';
+export * from '../models/users/user-dto/user.dto';
+export * from '../models/movie/movie-dto/movie.dto';
+export * from '../models/theatre/theatre-dto/theatre.dto';
+export * from '../models/booking/booking-dto/booking.dto';
