+++ src/models/movie/movie.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { AuthModule } from '../auth/auth.module';
+import { AuthModule } from '../../auth/auth.module';
 import { MovieController } from './movie.controller';
 import { MovieService } from './movie.service';
