+++ src/models/theatre/theatre.controller.ts
