+++ src/models/users/users.module.ts
@@ -1,5 +1,5 @@
 import { Module, forwardRef } from '@nestjs/common';
-import { AuthModule } from '../auth/auth.module';
+import { AuthModule } from '../../auth/auth.module';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
